chore: delete pages folder and move files to web folder
