+++ apps/web/app/editUpload/page.tsx
@@ -1,4 +1,4 @@
-import { EditUpload } from '@hilson/pages'
+import { EditUpload } from './EditUpload';
 
 export default function Index() {
   return (
