--- pages/project.json
@@ -1,9 +1,0 @@
-{
-  "name": "pages",
-  "$schema": "../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "pages/src",
-  "projectType": "library",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project pages --web",
-  "targets": {}
-}
