--- pages/src/index.ts
@@ -1,5 +1,0 @@
-export * from './JsonToCsv/JsonToCsv';
-export * from './BasicLayout/BasicLayout';
-export * from './EditUpload/EditUpload';
-export * from './Login/Login';
-export * from './Home/Home';
