+++ ui/src/index.ts
@@ -1,3 +1,2 @@
 export * from './BasicCheckbox/BasicCheckbox';
-export * from './BasicCheckbox/BasicCheckbox';
 export * from './BasicButton/BasicButton';
